refactor: consolidate overview tools into single flexible overview tool
Replace account-overview and project-overview with a unified overview tool:
- Remove: account-overview (no params), project-overview (projectId required)
- Add: overview (projectId optional)

The new tool provides:
- No projectId → Account overview (all projects + sections hierarchy)
- With projectId → Project overview (specific project + all tasks)

This consolidation reduces tool count from 20 to 19 (-1 tool) while maintaining
all existing functionality. The interface is intuitive with optional projectId
parameter, and output is consistently Markdown format in both modes.

Benefits:
- Simple, natural interface (optional projectId)
- Consistent output format (always Markdown)
- Preserves all functionality from both original tools
- Frees up tool budget for planned comment functionality
